Refactor (Training on undersampled instead of t-SNE)
Started training on the under-sampled data instead of the t-SNE data. I am unsure how good this model is at this point because its learning is unstable. I will have to plot the ROC and possibly find other parameters to evaluate the performance of the model. The t-SNE data seemed inconsistent and not suitable for training features. 

The learning rate was also increased to 0.5, which may be too fast for now. 

The accuracy so far is high and is included with the corresponding images.